segmenting code base
